+++ stub/resources/views/errors/400.blade.php
@@ -1,0 +1,3 @@
+<x-layout.error title="errors.bad_request" :link="true">
+
+</x-layout.error>